+++ src/providers/zenmux/utils.ts
@@ -1,0 +1,9 @@
+// Re-export utilities from HuggingFace provider since they're compatible
+export {
+	convertMessages,
+	convertTools,
+	isToolResultPart,
+	tryParseJSONObject,
+	validateRequest,
+	validateTools,
+} from "../huggingface/utils";